chore: split config into separate struct
