+++ h3/src/lib.rs
@@ -3,6 +3,7 @@
 #![allow(clippy::derive_partial_eq_without_eq)]
 
 pub mod client;
+mod config;
 pub mod error;
 pub mod ext;
 pub mod quic;
