+++ h3/src/server.rs
@@ -71,6 +71,7 @@ use quic::StreamId;
 use tokio::sync::mpsc;
 
 use crate::{
+    config::Config,
     connection::{self, ConnectionInner, ConnectionState, SharedStateRef},
     error::{Code, Error, ErrorLevel},
     frame::{FrameStream, FrameStreamError},
@@ -79,7 +80,6 @@ use crate::{
         frame::{Frame, PayloadLen},
         headers::Header,
         push::PushId,
-        varint::VarInt,
     },
     qpack,
     quic::{self, RecvDatagramExt, SendDatagramExt, SendStream as _},
@@ -506,48 +506,6 @@ where
 {
     fn drop(&mut self) {
         self.inner.close(Code::H3_NO_ERROR, "");
-    }
-}
-
-/// Configures the HTTP/3 connection
-#[derive(Debug, Clone, Copy)]
-pub struct Config {
-    /// Just like in HTTP/2, HTTP/3 also uses the concept of "grease"
-    /// to prevent potential interoperability issues in the future.
-    /// In HTTP/3, the concept of grease is used to ensure that the protocol can evolve
-    /// and accommodate future changes without breaking existing implementations.
-    pub send_grease: bool,
-    /// The MAX_FIELD_SECTION_SIZE in HTTP/3 refers to the maximum size of the dynamic table used in HPACK compression.
-    /// HPACK is the compression algorithm used in HTTP/3 to reduce the size of the header fields in HTTP requests and responses.
-
-    /// In HTTP/3, the MAX_FIELD_SECTION_SIZE is set to 12.
-    /// This means that the dynamic table used for HPACK compression can have a maximum size of 2^12 bytes, which is 4KB.
-    pub max_field_section_size: u64,
-
-    //=https://datatracker.ietf.org/doc/html/draft-ietf-webtrans-http3/#section-3.1
-    /// Sets `SETTINGS_ENABLE_WEBTRANSPORT` if enabled
-    pub enable_webtransport: bool,
-    /// https://www.rfc-editor.org/info/rfc8441 defines an extended CONNECT method in Section 4,
-    /// enabled by the SETTINGS_ENABLE_CONNECT_PROTOCOL parameter.
-    /// That parameter is only defined for HTTP/2.
-    /// for extended CONNECT in HTTP/3; instead, the SETTINGS_ENABLE_WEBTRANSPORT setting implies that an endpoint supports extended CONNECT.
-    pub enable_extended_connect: bool,
-    /// Enable HTTP Datagrams, see https://datatracker.ietf.org/doc/rfc9297/ for details
-    pub enable_datagram: bool,
-    /// The maximum number of concurrent streams that can be opened by the peer.
-    pub max_webtransport_sessions: u64,
-}
-
-impl Default for Config {
-    fn default() -> Self {
-        Self {
-            max_field_section_size: VarInt::MAX.0,
-            send_grease: true,
-            enable_webtransport: false,
-            enable_extended_connect: false,
-            enable_datagram: false,
-            max_webtransport_sessions: 0,
-        }
     }
 }
